sfs-54: Fixes cypress test

+++ finance-api/internal/service/process_failed_dd_collections.go
@@ -71,7 +71,7 @@ func (s *Service) validateFailedCollectionLine(ctx context.Context, logger *slog
 	_ = collectionTime.Scan(cd)
 	_ = processedDate.Scan(pd)
 	_ = processedTime.Scan(pd)
-	amount = int32(payment.Amount)
+	amount = payment.Amount
 	_ = notes.Scan(payment.ReasonCode)
 
 	ledgerCount, _ := s.store.CountDuplicateLedger(ctx, store.CountDuplicateLedgerParams{
